move styles file
